Move testing.base to testing.pipelines.base.
